Add files via upload
Edited one instance of int variable to double to allow for average score to display decimal numbers

+++ Classes/IndividualReview.java
@@ -8,7 +8,7 @@ public class IndividualReview {
     public double getReviewRating() {
         return this.reviewRating;
     }
-    public void setReviewRating(int reviewRating) {
+    public void setReviewRating(double reviewRating) {
         this.reviewRating = reviewRating;
     }
     public String getReviewDescription() {
